validaçao no restante

+++ src/main/java/com/ativ1/springboot/controllers/EmployeeController.java
@@ -2,8 +2,14 @@ package com.ativ1.springboot.controllers;
 
 import com.ativ1.springboot.domain.Employee;
 import com.ativ1.springboot.repository.EmployeeRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
+import java.util.Map;
 
 //Endopoint DE FUNCIONARIOS
 @RestController
@@ -16,8 +22,9 @@ public class EmployeeController {
     }
 
     @PostMapping
-    public Employee Create(@RequestBody Employee employee) {
-        return repository.save(employee);
+    public ResponseEntity<Employee> Create(@RequestBody Employee employee) {
+        var employeeCreated = repository.save(employee);
+        return ResponseEntity.status(201).body(employeeCreated);
     }
 
     @GetMapping
@@ -26,8 +33,10 @@ public class EmployeeController {
     }
 
     @GetMapping("/{id}")
-    public Employee ListById(@PathVariable Long id) {
-        return repository.findById(id).orElse(null);
+    public ResponseEntity<?> ListById(@PathVariable Long id) {
+        var employee =  repository.findById(id);
+        if(employee.isPresent()) return ResponseEntity.ok(employee.get());
+        return ResponseEntity.status(404).body(Map.of("error", "Empregado não encontrado."));
     }
 
     @PutMapping("/{id}")
@@ -37,11 +46,19 @@ public class EmployeeController {
             employee.setPosition(newEmployee.getPosition());
             employee.setSalary(newEmployee.getSalary());
             return repository.save(employee);
-        }).orElse(null);
+        }).orElseThrow(()->  new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
     @DeleteMapping("/{id}")
-    public void Delete(@PathVariable Long id) {
-        repository.deleteById(id);
-    }
+    public ResponseEntity<?> Delete(@PathVariable Long id) {
+        var employee = repository.findById(id);
+        if (employee.isPresent()) {
+            repository.deleteById(id);
+            return ResponseEntity.status(200)
+                    .body(Map.of("message", "Empregado deletado com sucesso."));
+        }
+        return ResponseEntity
+                .status(404)
+                .body(Map.of("error", "Empregado não encontrado."));}
+
 }
